% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-imap.R, R/future-walk.R
\name{future_imap}
\alias{future_imap}
\alias{future_imap_chr}
\alias{future_imap_dbl}
\alias{future_imap_int}
\alias{future_imap_lgl}
\alias{future_imap_vec}
\alias{future_imap_dfr}
\alias{future_imap_dfc}
\alias{future_iwalk}
\title{Apply a function to each element of a vector, and its index via futures}
\usage{
future_imap(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_chr(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_dbl(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_int(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_lgl(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_vec(
  .x,
  .f,
  ...,
  .ptype = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_dfr(
  .x,
  .f,
  ...,
  .id = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_imap_dfc(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_iwalk(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, specified in one of the following ways:
\itemize{
\item A named function, e.g. \code{paste}.
\item An anonymous function, e.g. \verb{\\(x, idx) x + idx} or
\code{function(x, idx) x + idx}.
\item A formula, e.g. \code{~ .x + .y}. Use \code{.x} to refer to the current element and
\code{.y} to refer to the current index. No longer recommended.
}}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> future_map(f, 1, 2, collapse = ",")
# do:
x |> future_map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{.options}{The \code{future} specific options to use with the workers. This
must be the result from a call to \code{\link[=furrr_options]{furrr_options()}}.}

\item{.env_globals}{The environment to look for globals required by \code{.x} and
\code{...}. Globals required by \code{.f} are looked up in the function environment
of \code{.f}.}

\item{.progress}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Either \code{TRUE} or \code{FALSE}.

TODO!(write this): See \code{vignette("progress")} for examples.

If \code{TRUE}, a \code{progressr::progressor()} is automatically created
with as many steps as there are elements to iterate over. After each
call to \code{.f}, progression is \emph{signaled} back to the main process.

Note that \code{.progress = TRUE} is not enough to make progress notifications
\emph{appear}. You must also either wrap the furrr expression in
\code{\link[progressr:with_progress]{progressr::with_progress()}} or set \code{\link[progressr:handlers]{progressr::handlers(global = TRUE)}} at the top of your script (this enables
\emph{all} progress notifications, which is recommended for interative work).

To customize the progress notification, set an alternative progressr
handler, such as
\code{\link[progressr:handler_cli]{progressr::handlers(progressr::handler_cli())}}
or
\code{\link[progressr:handler_rstudio]{progressr::handlers(progressr::handler_rstudio())}}.

Note that progress notifications are not free, as they require
communicating a small amount of data back to the main worker after each
call to \code{.f}. It is recommended to only use them with long running tasks.

All core \code{plan()} types are supported, including \code{sequential}, \code{cluster},
\code{multisession}, and \code{multicore}. Additionally, \code{future.callr::callr} and
\code{future.mirai::mirai_multisession} are also known to be supported.}

\item{.ptype}{If \code{NULL}, the default, the output type is the common type of
the elements of the result. Otherwise, supply a "prototype" giving the
desired type of output.}

\item{.id}{Either a string or \code{NULL}. If a string, the output will contain
a variable with that name, storing either the name (if \code{.x} is named) or
the index (if \code{.x} is unnamed) of the input. If \code{NULL}, the default, no
variable will be created.

Only applies to \verb{_dfr} variant.}
}
\value{
A vector the same length as .x.
}
\description{
These functions work the same as \code{\link[purrr:imap]{purrr::imap()}} functions,
but allow you to map in parallel.
}
\examples{
\donttest{plan(multisession, workers = 2)}

future_imap_chr(sample(10), ~ paste0(.y, ": ", .x))

\dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
