% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-map2.R, R/future-pmap.R, R/future-walk.R
\name{future_map2}
\alias{future_map2}
\alias{future_map2_chr}
\alias{future_map2_dbl}
\alias{future_map2_int}
\alias{future_map2_lgl}
\alias{future_map2_vec}
\alias{future_map2_dfr}
\alias{future_map2_dfc}
\alias{future_pmap}
\alias{future_pmap_chr}
\alias{future_pmap_dbl}
\alias{future_pmap_int}
\alias{future_pmap_lgl}
\alias{future_pmap_vec}
\alias{future_pmap_dfr}
\alias{future_pmap_dfc}
\alias{future_walk2}
\alias{future_pwalk}
\title{Map over multiple inputs simultaneously via futures}
\usage{
future_map2(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_chr(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_dbl(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_int(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_lgl(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_vec(
  .x,
  .y,
  .f,
  ...,
  .ptype = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_dfr(
  .x,
  .y,
  .f,
  ...,
  .id = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map2_dfc(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_chr(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_dbl(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_int(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_lgl(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_vec(
  .l,
  .f,
  ...,
  .ptype = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_dfr(
  .l,
  .f,
  ...,
  .id = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pmap_dfc(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_walk2(
  .x,
  .y,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_pwalk(
  .l,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)
}
\arguments{
\item{.x, .y}{A pair of vectors, usually the same length. If not, a vector
of length 1 will be recycled to the length of the other.}

\item{.f}{A function, specified in one of the following ways:
\itemize{
\item A named function.
\item An anonymous function, e.g. \verb{\\(x, y) x + y} or \code{function(x, y) x + y}.
\item A formula, e.g. \code{~ .x + .y}. Use \code{.x} to refer to the current
element of \code{x} and \code{.y} to refer to the current element of \code{y}.
No longer recommended.
}}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> future_map(f, 1, 2, collapse = ",")
# do:
x |> future_map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{.options}{The \code{future} specific options to use with the workers. This
must be the result from a call to \code{\link[=furrr_options]{furrr_options()}}.}

\item{.env_globals}{The environment to look for globals required by \code{.x} and
\code{...}. Globals required by \code{.f} are looked up in the function environment
of \code{.f}.}

\item{.progress}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Either \code{TRUE} or \code{FALSE}.

TODO!(write this): See \code{vignette("progress")} for examples.

If \code{TRUE}, a \code{progressr::progressor()} is automatically created
with as many steps as there are elements to iterate over. After each
call to \code{.f}, progression is \emph{signaled} back to the main process.

Note that \code{.progress = TRUE} is not enough to make progress notifications
\emph{appear}. You must also either wrap the furrr expression in
\code{\link[progressr:with_progress]{progressr::with_progress()}} or set \code{\link[progressr:handlers]{progressr::handlers(global = TRUE)}} at the top of your script (this enables
\emph{all} progress notifications, which is recommended for interative work).

To customize the progress notification, set an alternative progressr
handler, such as
\code{\link[progressr:handler_cli]{progressr::handlers(progressr::handler_cli())}}
or
\code{\link[progressr:handler_rstudio]{progressr::handlers(progressr::handler_rstudio())}}.

Note that progress notifications are not free, as they require
communicating a small amount of data back to the main worker after each
call to \code{.f}. It is recommended to only use them with long running tasks.

All core \code{plan()} types are supported, including \code{sequential}, \code{cluster},
\code{multisession}, and \code{multicore}. Additionally, \code{future.callr::callr} and
\code{future.mirai::mirai_multisession} are also known to be supported.}

\item{.ptype}{If \code{NULL}, the default, the output type is the common type of
the elements of the result. Otherwise, supply a "prototype" giving the
desired type of output.}

\item{.id}{Either a string or \code{NULL}. If a string, the output will contain
a variable with that name, storing either the name (if \code{.x} is named) or
the index (if \code{.x} is unnamed) of the input. If \code{NULL}, the default, no
variable will be created.

Only applies to \verb{_dfr} variant.}

\item{.l}{A list of vectors. The length of \code{.l} determines the number of
arguments that \code{.f} will be called with. Arguments will be supply by
position if unnamed, and by name if named.

Vectors of length 1 will be recycled to any length; all other elements
must be have the same length.

A data frame is an important special case of \code{.l}. It will cause \code{.f}
to be called once for each row.}
}
\value{
An atomic vector, list, or data frame, depending on the suffix.
Atomic vectors and lists will be named if \code{.x} or the first element of \code{.l}
is named.

If all input is length 0, the output will be length 0. If any input is
length 1, it will be recycled to the length of the longest.
}
\description{
These functions work the same as \code{\link[purrr:map2]{purrr::map2()}} and its variants,
but allow you to map in parallel. Note that "parallel" as described in purrr
is just saying that you are working with multiple inputs, and parallel in
this case means that you can work on multiple inputs and process them all in
parallel as well.
}
\examples{
\donttest{plan(multisession, workers = 2)}

x <- list(1, 10, 100)
y <- list(1, 2, 3)
z <- list(5, 50, 500)

future_map2(x, y, ~ .x + .y)

# Split into pieces, fit model to each piece, then predict
by_cyl <- split(mtcars, mtcars$cyl)
mods <- future_map(by_cyl, ~ lm(mpg ~ wt, data = .))
future_map2(mods, by_cyl, predict)

future_pmap(list(x, y, z), sum)

# Matching arguments by position
future_pmap(list(x, y, z), function(a, b ,c) a / (b + c))

# Vectorizing a function over multiple arguments
df <- data.frame(
  x = c("apple", "banana", "cherry"),
  pattern = c("p", "n", "h"),
  replacement = c("x", "f", "q"),
  stringsAsFactors = FALSE
)

future_pmap(df, gsub)
future_pmap_chr(df, gsub)

\dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
