% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-modify.R
\name{future_modify}
\alias{future_modify}
\alias{future_modify_at}
\alias{future_modify_if}
\title{Modify elements selectively via futures}
\usage{
future_modify(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_modify_at(
  .x,
  .at,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_modify_if(
  .x,
  .p,
  .f,
  ...,
  .else = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)
}
\arguments{
\item{.x}{A vector.}

\item{.f}{A function specified in the same way as the corresponding map
function.}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> future_map(f, 1, 2, collapse = ",")
# do:
x |> future_map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{.options}{The \code{future} specific options to use with the workers. This
must be the result from a call to \code{\link[=furrr_options]{furrr_options()}}.}

\item{.env_globals}{The environment to look for globals required by \code{.x} and
\code{...}. Globals required by \code{.f} are looked up in the function environment
of \code{.f}.}

\item{.progress}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Either \code{TRUE} or \code{FALSE}.

TODO!(write this): See \code{vignette("progress")} for examples.

If \code{TRUE}, a \code{progressr::progressor()} is automatically created
with as many steps as there are elements to iterate over. After each
call to \code{.f}, progression is \emph{signaled} back to the main process.

Note that \code{.progress = TRUE} is not enough to make progress notifications
\emph{appear}. You must also either wrap the furrr expression in
\code{\link[progressr:with_progress]{progressr::with_progress()}} or set \code{\link[progressr:handlers]{progressr::handlers(global = TRUE)}} at the top of your script (this enables
\emph{all} progress notifications, which is recommended for interative work).

To customize the progress notification, set an alternative progressr
handler, such as
\code{\link[progressr:handler_cli]{progressr::handlers(progressr::handler_cli())}}
or
\code{\link[progressr:handler_rstudio]{progressr::handlers(progressr::handler_rstudio())}}.

Note that progress notifications are not free, as they require
communicating a small amount of data back to the main worker after each
call to \code{.f}. It is recommended to only use them with long running tasks.

All core \code{plan()} types are supported, including \code{sequential}, \code{cluster},
\code{multisession}, and \code{multicore}. Additionally, \code{future.callr::callr} and
\code{future.mirai::mirai_multisession} are also known to be supported.}

\item{.at}{A logical, integer, or character vector giving the elements
to select. Alternatively, a function that takes a vector of names,
and returns a logical, integer, or character vector of elements to select.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}: if the tidyselect package is
installed, you can use \code{vars()} and tidyselect helpers to select
elements.}

\item{.p}{A single predicate function, a formula describing such a
predicate function, or a logical vector of the same length as \code{.x}.
Alternatively, if the elements of \code{.x} are themselves lists of
objects, a string indicating the name of a logical element in the
inner lists. Only those elements where \code{.p} evaluates to
\code{TRUE} will be modified.}

\item{.else}{A function applied to elements of \code{.x} for which \code{.p}
returns \code{FALSE}.}
}
\value{
An object the same class as \code{.x}
}
\description{
These functions work the same as \code{\link[purrr:modify]{purrr::modify()}} functions, but
allow you to modify in parallel.
}
\details{
From purrr:

Since the transformation can alter the structure of the input;
it's your responsibility to ensure that the transformation produces a valid
output. For example, if you're modifying a data frame, \code{.f} must preserve the
length of the input.
}
\examples{
\donttest{plan(multisession, workers = 2)}

# Convert each col to character, in parallel
future_modify(mtcars, as.character)

iris |>
 future_modify_if(is.factor, as.character) |>
 str()

mtcars |>
  future_modify_at(c(1, 4, 5), as.character) |>
  str()

\dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
