% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plan}
\alias{tweak}
\alias{makeClusterPSOCK}
\alias{cluster}
\alias{multicore}
\alias{multisession}
\alias{multiprocess}
\alias{remote}
\alias{sequential}
\alias{transparent}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{future}{\code{\link[future]{cluster}}, \code{\link[future]{makeClusterPSOCK}}, \code{\link[future]{multicore}}, \code{\link[future]{multiprocess}}, \code{\link[future]{multisession}}, \code{\link[future]{plan}}, \code{\link[future]{remote}}, \code{\link[future]{sequential}}, \code{\link[future:sequential]{transparent}}, \code{\link[future]{tweak}}}
}}

