% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furrr-package.R
\docType{package}
\name{furrr-package}
\alias{furrr}
\alias{furrr-package}
\title{furrr: Apply Mapping Functions in Parallel using Futures}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implementations of the family of map() functions from 'purrr' that can be resolved using any 'future'-supported backend, e.g. parallel on the local machine or distributed on a compute cluster.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/furrr}
  \item \url{https://furrr.futureverse.org/}
  \item Report bugs at \url{https://github.com/DavisVaughan/furrr/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Authors:
\itemize{
  \item Matt Dancho \email{mdancho@business-science.io}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
