% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future-map.R, R/future-walk.R
\name{future_map}
\alias{future_map}
\alias{future_map_chr}
\alias{future_map_dbl}
\alias{future_map_int}
\alias{future_map_lgl}
\alias{future_map_vec}
\alias{future_map_dfr}
\alias{future_map_dfc}
\alias{future_walk}
\title{Apply a function to each element of a vector via futures}
\usage{
future_map(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_chr(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_dbl(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_int(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_lgl(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_vec(
  .x,
  .f,
  ...,
  .ptype = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_dfr(
  .x,
  .f,
  ...,
  .id = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_map_dfc(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)

future_walk(
  .x,
  .f,
  ...,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = FALSE
)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, specified in one of the following ways:
\itemize{
\item A named function, e.g. \code{mean}.
\item An anonymous function, e.g. \verb{\\(x) x + 1} or \code{function(x) x + 1}.
\item A formula, e.g. \code{~ .x + 1}. Use \code{.x} to refer to the first
argument. No longer recommended.
\item A string, integer, or list, e.g. \code{"idx"}, \code{1}, or \code{list("idx", 1)} which
are shorthand for \verb{\\(x) pluck(x, "idx")}, \verb{\\(x) pluck(x, 1)}, and
\verb{\\(x) pluck(x, "idx", 1)} respectively. Optionally supply \code{.default} to
set a default value if the indexed element is \code{NULL} or does not exist.
}}

\item{...}{Additional arguments passed on to the mapped function.

We now generally recommend against using \code{...} to pass additional
(constant) arguments to \code{.f}. Instead use a shorthand anonymous function:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{# Instead of
x |> future_map(f, 1, 2, collapse = ",")
# do:
x |> future_map(\\(x) f(x, 1, 2, collapse = ","))
}\if{html}{\out{</div>}}

This makes it easier to understand which arguments belong to which
function and will tend to yield better error messages.}

\item{.options}{The \code{future} specific options to use with the workers. This
must be the result from a call to \code{\link[=furrr_options]{furrr_options()}}.}

\item{.env_globals}{The environment to look for globals required by \code{.x} and
\code{...}. Globals required by \code{.f} are looked up in the function environment
of \code{.f}.}

\item{.progress}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Either \code{TRUE} or \code{FALSE}.

TODO!(write this): See \code{vignette("progress")} for examples.

If \code{TRUE}, a \code{progressr::progressor()} is automatically created
with as many steps as there are elements to iterate over. After each
call to \code{.f}, progression is \emph{signaled} back to the main process.

Note that \code{.progress = TRUE} is not enough to make progress notifications
\emph{appear}. You must also either wrap the furrr expression in
\code{\link[progressr:with_progress]{progressr::with_progress()}} or set \code{\link[progressr:handlers]{progressr::handlers(global = TRUE)}} at the top of your script (this enables
\emph{all} progress notifications, which is recommended for interative work).

To customize the progress notification, set an alternative progressr
handler, such as
\code{\link[progressr:handler_cli]{progressr::handlers(progressr::handler_cli())}}
or
\code{\link[progressr:handler_rstudio]{progressr::handlers(progressr::handler_rstudio())}}.

Note that progress notifications are not free, as they require
communicating a small amount of data back to the main worker after each
call to \code{.f}. It is recommended to only use them with long running tasks.

All core \code{plan()} types are supported, including \code{sequential}, \code{cluster},
\code{multisession}, and \code{multicore}. Additionally, \code{future.callr::callr} and
\code{future.mirai::mirai_multisession} are also known to be supported.}

\item{.ptype}{If \code{NULL}, the default, the output type is the common type of
the elements of the result. Otherwise, supply a "prototype" giving the
desired type of output.}

\item{.id}{Either a string or \code{NULL}. If a string, the output will contain
a variable with that name, storing either the name (if \code{.x} is named) or
the index (if \code{.x} is unnamed) of the input. If \code{NULL}, the default, no
variable will be created.

Only applies to \verb{_dfr} variant.}
}
\value{
All functions return a vector the same length as \code{.x}.
\itemize{
\item \code{\link[=future_map]{future_map()}} returns a list
\item \code{\link[=future_map_lgl]{future_map_lgl()}} a logical vector
\item \code{\link[=future_map_int]{future_map_int()}} an integer vector
\item \code{\link[=future_map_dbl]{future_map_dbl()}} a double vector
\item \code{\link[=future_map_chr]{future_map_chr()}} a character vector
}

The output of \code{.f} will be automatically typed upwards, e.g. logical ->
integer -> double -> character.
}
\description{
These functions work the same as \code{\link[purrr:map]{purrr::map()}} and its variants, but
allow you to map in parallel.
}
\examples{
\donttest{plan(multisession, workers = 2)}

1:10 |>
  future_map(rnorm, n = 10, .options = furrr_options(seed = 123)) |>
  future_map_dbl(mean)

# If each element of the output is a data frame, use
# `future_map_dfr()` to row-bind them together:
mtcars |>
  split(mtcars$cyl) |>
  future_map(~ lm(mpg ~ wt, data = .x)) |>
  future_map_dfr(~ as.data.frame(t(as.matrix(coef(.)))))


# You can be explicit about what gets exported to the workers.
# To see this, use multisession (not multicore as the forked workers
# still have access to this environment)
\donttest{plan(multisession)}
x <- 1
y <- 2

# This will fail, y is not exported (no black magic occurs)
try(future_map(1, ~y, .options = furrr_options(globals = "x")))

# y is exported
future_map(1, ~y, .options = furrr_options(globals = "y"))

\dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
}
}
