% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furrr-options.R
\name{future_options}
\alias{future_options}
\title{Deprecated furrr options}
\usage{
future_options(globals = TRUE, packages = NULL, seed = FALSE, scheduling = 1)
}
\arguments{
\item{globals}{A logical, a character vector, a named list, or \code{NULL} for
controlling how globals are handled. For details, see the
\verb{Global variables} section below.}

\item{packages}{A character vector, or \code{NULL}. If supplied, this specifies
packages that are guaranteed to be attached in the R environment where the
future is evaluated.}

\item{seed}{A logical, an integer of length \code{1} or \code{7}, a list of
\code{length(.x)} with pre-generated random seeds, or \code{NULL}. For details, see
the \verb{Reproducible random number generation (RNG)} section below.}

\item{scheduling}{A single integer, logical, or \code{Inf}. This argument
controls the average number of futures ("chunks") per worker.
\itemize{
\item If \code{0}, then a single future is used to process all elements of \code{.x}.
\item If \code{1} or \code{TRUE}, then one future per worker is used.
\item If \code{2}, then each worker will process two futures (provided there
are enough elements in \code{.x}).
\item If \code{Inf} or \code{FALSE}, then one future per element of \code{.x} is used.
}

This argument is only used if \code{chunk_size} is \code{NULL}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

As of furrr 0.3.0, \code{future_options()} is defunct in favor of
\code{\link[=furrr_options]{furrr_options()}}.
}
\examples{
try(future_options())
}
\keyword{internal}
